Add doChangeThumbnailSize.js
